api/user/requirements.txt
Flask==2.0.3
google-cloud-firestore==2.1.0
gunicorn==20.1.0
Werkzeug==2.0.3